Refactors to address review comment
