+++ pinot-common/src/main/java/org/apache/pinot/common/utils/RoaringBitmapUtils.java
@@ -21,6 +21,7 @@ package org.apache.pinot.common.utils;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import org.roaringbitmap.ImmutableBitmapDataProvider;
+import org.roaringbitmap.IntIterator;
 import org.roaringbitmap.RoaringBitmap;
 
 
@@ -51,5 +52,38 @@ public class RoaringBitmapUtils {
       throw new RuntimeException("Caught exception while deserializing RoaringBitmap", e);
     }
     return bitmap;
+  }
+
+  /**
+   * Iterates over the ranges of unset bits and calls the consumer for each range. This is more performant to
+   * alternatives like calling {@link RoaringBitmap#contains(int)} in a loop or cloning and flipping the bitmap before
+   * iterating, especially for sparse bitmaps.
+   * @param nullIndexIterator an int iterator that returns values in ascending order whose min value is 0.
+   */
+  public static void forEachUnset(int length, IntIterator nullIndexIterator, BatchConsumer consumer) {
+    int prev = 0;
+    while (nullIndexIterator.hasNext() && prev < length) {
+      int nextNull = Math.min(nullIndexIterator.next(), length);
+      if (nextNull > prev) {
+        consumer.consume(prev, nextNull);
+      }
+      prev = nextNull + 1;
+    }
+    if (prev < length) {
+      consumer.consume(prev, length);
+    }
+  }
+
+  /**
+   * A consumer that is being used to consume batch of indexes.
+   */
+  @FunctionalInterface
+  public interface BatchConsumer {
+    /**
+     * Consumes a batch of indexes.
+     * @param fromInclusive the start index (inclusive)
+     * @param toExclusive the end index (exclusive)
+     */
+    void consume(int fromInclusive, int toExclusive);
   }
 }
