+++ pinot-core/src/main/java/org/apache/pinot/core/operator/transform/function/CastTransformFunction.java
@@ -22,12 +22,14 @@ import com.google.common.base.Preconditions;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
+import org.apache.pinot.common.utils.RoaringBitmapUtils;
 import org.apache.pinot.core.operator.ColumnContext;
 import org.apache.pinot.core.operator.blocks.ValueBlock;
 import org.apache.pinot.core.operator.transform.TransformResultMetadata;
 import org.apache.pinot.spi.data.FieldSpec.DataType;
 import org.apache.pinot.spi.utils.ArrayCopyUtils;
-import org.roaringbitmap.IntConsumer;
+import org.apache.pinot.spi.utils.TimestampUtils;
+import org.roaringbitmap.RoaringBitmap;
 
 
 public class CastTransformFunction extends BaseTransformFunction {
@@ -192,12 +194,19 @@ public class CastTransformFunction extends BaseTransformFunction {
       int length = valueBlock.getNumDocs();
       initLongValuesSV(length);
       String[] stringValues = _transformFunction.transformToStringValuesSV(valueBlock);
-      // Null string values can't be converted to timestamp, so replace with actual null to allow the util to skip
-      // those values
-      if (_transformFunction.getNullBitmap(valueBlock) != null) {
-        _transformFunction.getNullBitmap(valueBlock).forEach((IntConsumer) index -> stringValues[index] = null);
+      RoaringBitmap nullBitmap = _transformFunction.getNullBitmap(valueBlock);
+      if (nullBitmap != null) {
+        // Null string values can't be converted to valid timestamps, so we skip over those values.
+        // Avoid using RoaringBitmap::contains API in a loop due to poor performance.
+        // Avoid cloning + flipping the null bitmap to reduce allocation.
+        RoaringBitmapUtils.forEachUnset(length, nullBitmap.getIntIterator(), (from, to) -> {
+          for (int i = from; i < to; i++) {
+            _longValuesSV[i] = TimestampUtils.toMillisSinceEpoch(stringValues[i]);
+          }
+        });
+      } else {
+        ArrayCopyUtils.copyToTimestamp(stringValues, _longValuesSV, length);
       }
-      ArrayCopyUtils.copyToTimestamp(stringValues, _longValuesSV, length);
       return _longValuesSV;
     } else {
       return _transformFunction.transformToLongValuesSV(valueBlock);
