+++ pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/LastLongValueWithTimeAggregationFunction.java
@@ -21,6 +21,7 @@ package org.apache.pinot.core.query.aggregation.function;
 import org.apache.pinot.common.CustomObject;
 import org.apache.pinot.common.request.context.ExpressionContext;
 import org.apache.pinot.common.utils.DataSchema.ColumnDataType;
+import org.apache.pinot.common.utils.RoaringBitmapUtils;
 import org.apache.pinot.core.common.BlockValSet;
 import org.apache.pinot.core.common.ObjectSerDeUtils;
 import org.apache.pinot.core.query.aggregation.groupby.GroupByResultHolder;
@@ -69,7 +70,7 @@ public class LastLongValueWithTimeAggregationFunction extends LastWithTimeAggreg
     long[] timeValues = timeValSet.getLongValuesSV();
 
     IntIterator nullIdxIterator = orNullIterator(blockValSet, timeValSet);
-    forEachNotNull(length, nullIdxIterator, (from, to) -> {
+    RoaringBitmapUtils.forEachUnset(length, nullIdxIterator, (from, to) -> {
       for (int i = from; i < to; i++) {
         long data = longValues[i];
         long time = timeValues[i];
@@ -85,7 +86,7 @@ public class LastLongValueWithTimeAggregationFunction extends LastWithTimeAggreg
     long[] timeValues = timeValSet.getLongValuesSV();
 
     IntIterator nullIdxIterator = orNullIterator(blockValSet, timeValSet);
-    forEachNotNull(length, nullIdxIterator, (from, to) -> {
+    RoaringBitmapUtils.forEachUnset(length, nullIdxIterator, (from, to) -> {
       for (int i = from; i < to; i++) {
         long value = longValues[i];
         long time = timeValues[i];
